requirements.txt
bcrypt==3.1.7
Django==3.1.3
django-cors-headers==3.10.0
django-extensions==3.1.5
djangorestframework==3.12.2
gunicorn==20.1.0
social-auth-core==4.1.0
sqlparse==0.4.1
stack-data==0.2.0
urllib3==1.25.8
virtualenv==20.0.17
Werkzeug==1.0.1
asgiref==3.4.1
pip==21.1.3
pytz==2021.1
setuptools==56.0.0
bs4==0.0.1
requests==2.7.0
json==0.3.0